django_project/django_app/serializers/match_prediction_serializer.py
from rest_framework import serializers
from ..models import MatchPrediction

class MatchPredictionSerializer(serializers.ModelSerializer):
    class Meta:
        model = MatchPrediction
        fields = '__all__'
