django_project/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
mysqlclient==2.2.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
kafka-python==2.0.2
pyspark==3.5.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
tweepy==4.14.0
